Refine add feed

+++ add-feed.sh
@@ -1,10 +1,4 @@
 #!/bin/sh -e
-# check if feed exists
-
-if grep -q "ekkog" /etc/opkg/customfeeds.conf; then
-    echo "Feed already added"
-    exit 0
-fi
 
 . /etc/openwrt_release 
 DISTRIB_RELEASE=$(echo $DISTRIB_RELEASE | awk -F. '{print $1"."$2}')
@@ -13,19 +7,24 @@ if [ "$DISTRIB_RELEASE" = "snapshot" ]; then
 elif [ "$ISTRIB_RELEASE" != "21.02" ] && [ "$DISTRIB_RELEASE" != "22.03" ]; then
     DISTRIB_RELEASE="21.02"
 fi
-feed="
-src/gz ekkog_packages https://github.com/ekkog/openwrt-dist/raw/packages/${DISTRIB_ARCH}-${DISTRIB_RELEASE}
-src/gz ekkog_luci https://github.com/ekkog/openwrt-dist/raw/luci/${DISTRIB_RELEASE}
-"
 
-echo "Add feed..."
+if grep -q "ekkog" /etc/opkg/customfeeds.conf; then
+    echo "Feed already added"
+else
+    feed="
+    src/gz ekkog_packages https://github.com/ekkog/openwrt-dist/raw/packages/${DISTRIB_ARCH}-${DISTRIB_RELEASE}
+    src/gz ekkog_luci https://github.com/ekkog/openwrt-dist/raw/luci/${DISTRIB_RELEASE}
+    "
+    echo "Add feed..."
+    echo "$feed" >> /etc/opkg/customfeeds.conf
+fi
 
-echo "$feed" >> /etc/opkg/customfeeds.conf
-
-echo "Add key..."
-wget https://github.com/ekkog/openwrt-dist/raw/master/cd5844109a8e9dda
-opkg-key add cd5844109a8e9dda
-rm cd5844109a8e9dda
+if [ ! -f /etc/opkg/keys/cd5844109a8e9dda ]; then
+    echo "Add key..."
+    wget https://github.com/ekkog/openwrt-dist/raw/master/cd5844109a8e9dda
+    opkg-key add cd5844109a8e9dda
+    rm cd5844109a8e9dda
+fi
 
 echo "Update feeds..."
